Adding the deterministicOutput option.
It is now possible to specify the --deterministic-output command line
option or the deterministicOutput grunt option to produce a file that is
easier to compare with other files of the same kind (with classic diff
tools): the resulting json file contains meaningful ids instead of
sequential ones and json keys are sorted so that the order in which files
are compared does not impact the output.

This option is used to better test at-diff.